pyproject.toml
[project]
name = "sqlnosql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faker>=37.4.0",
    "markdown2>=2.5.3",
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "pymongo>=4.13.1",
    "python-dotenv>=1.1.0",
    "weasyprint>=65.1",
]
